Make redirects e2e test aware
Since the Karma e2e test runner proxies requests on behalf of an iframe
for test execution, we have to ensure that the iframe doesn't get
redirected to the Rails server's domain. Otherwise, we get cross domain
security errors when Karma tries to reach into the iframe.

The next step will be to create a middleware to handle this
automatically rather than sprinkling our application code with e2e test
aware logic.